测试静态文件

+++ src/main/java/com/pichub/hello/web/BelikeController.java
@@ -28,7 +28,7 @@ public class BelikeController {
     @ResponseBody
     public String checkBelike(long pictureId, HttpServletRequest request)throws Exception
     {
-        return "{" + userService.belikeCheck(User.getCurrentUser(request).getUserId().longValue(),pictureId) + "}";
+        return "{" + userService.belikeCheck(User.getCurrentUser(request).getUserId(),pictureId) + "}";
     }
 
     @RequestMapping(value = "/belike/reverseState")
